% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_ueb.R
\name{barplot_ueb}
\alias{barplot_ueb}
\title{A barplot_ueb Function}
\usage{
barplot_ueb(
  y,
  group = NULL,
  frml = NULL,
  dat,
  las = 0,
  title.plot = NULL,
  sub.plot = NULL,
  cex.lab = 1,
  cex.main = 1,
  cex.lg = 1,
  cex.n = 0.8,
  ylab = "",
  title.lg = FALSE,
  do.test = FALSE,
  show.lg = FALSE,
  show.freq = TRUE
)
}
\arguments{
\item{y}{name factor vector of data values.}

\item{group}{name factor vector. Default value is NULL.}

\item{frml}{Right side of ~ must have the terms in an additive way, and left side of ~ must contain the name of the grouping variable or can be left in blank (in this latter case descriptives for whole sample are calculated and no test is performed).}

\item{dat}{matrix or data frame containing the variables in the formula.}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. 0: always parallel to the axis [default],
1: always horizontal, 2:always perpendicular to the axis, 3: always vertical.}

\item{title.plot}{a main title for the plot}

\item{sub.plot}{a sub title for the plot}

\item{cex.lab}{expansion factor for axis names (bar labels). (size x labels)}

\item{cex.main}{expansion factor for main names (size main)}

\item{cex.lg}{expansion factor for legend names (size legend)}

\item{cex.n}{expansion factor for size n}

\item{ylab}{a title for the y axis}

\item{do.test}{logical value si se quiere realizar test Chi cuadrado SIN correcci칩n de yates.}

\item{show.lg}{TRUE o FALSE indica si se muestra la leyenda. Por defecto FALSE.}

\item{show.freq}{TRUE o FALSE indica si se muestran las frecuencias. Por defecto TRUE}

\item{title.leg}{TRUE o FALSE indica si se muestra el titulo de la leyenda. Por defecto FALSE.}
}
\description{
Barplot univariante o bivariante mostrando frecuencias relativas. En el caso
del univariante es posible a침adir las frecuencias absolutas. Para el an치lisis
bivariante es posible realizar test Chi cuadrado.
}
\examples{
# df <- data.frame(sex = c(sample(c('Male Male', 'Female Female'), 500, replace = TRUE, prob = c(.2,.8) ),
# sample(c('Male Male', 'Female Female'), 500, replace = TRUE, prob = c(.4,.6) )),
# grup =  c( rep('Casos', 500),rep('Control', 500)  ))
# barplot_ueb(y = "grup", dat = df)
# barplot_ueb(y = "sex",group = "grup", dat = df, cex.lab = 0.8, do.test = TRUE)
# El mateix cridant la funci칩 amb formula
# barplot_ueb(frml = ~ sex, dat = df)
# barplot_ueb(frml =  grup ~ sex, dat = df, cex.lab = 0.8, do.test = TRUE)
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{barplot}
\keyword{descriptive}
\keyword{plots}
