% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_bw.R
\name{boxplot_bw}
\alias{boxplot_bw}
\title{A boxplot_bw Function}
\usage{
boxplot_bw(y, group = NULL, dat, las = 0, title.plot = NULL,
  ylim.plot = NULL, cex.lab = 1)
}
\arguments{
\item{y}{nombre de la variable numerica}

\item{group}{nombre de la variable factor. Default value is NULL.}

\item{dat}{data frame que contiene las variables a graficar.}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. 0: always parallel to the axis [default],
1: always horizontal, 2:always perpendicular to the axis, 3: always vertical.}

\item{title.plot}{a main title for the plot}

\item{ylim.plot}{is vector which contains lower and upper limits which are to appear on the y axes.}

\item{title}{titulo del grafico}
}
\description{
Boxplot incluyendo puntos individuales
}
\examples{
df <- data.frame(runif = c(runif(100, min = -3, max = 3),rep(0,25)), rnorm = c(rnorm(100),rep(0,25)) )
boxplot_bw(dat = df, y = 'rnorm' )
boxplot_bw(dat = mtc_bis, y = 'qsec' )
boxplot_bw(dat = mtc_bis, y = 'qsec', group = 'gear')
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{boxplot}
\keyword{descriptive}
\keyword{plots}
