% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.test2.R
\name{chisq.test2}
\alias{chisq.test2}
\title{A chisq.test2 Function}
\usage{
chisq.test2(obj, chisq.test.perm)
}
\description{
chisq.test2 para no aplicar corrección de Yates en el paquete compareGroups.
}
\keyword{chisquare}
\keyword{comparegroups}
\keyword{correction}
\keyword{test}
\keyword{yates}
