% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dictionary.R
\name{create_dictionary}
\alias{create_dictionary}
\title{A create_dictionary Function}
\usage{
create_dictionary(var, name.var = NULL)
}
\arguments{
\item{var}{object to be coerced or tested. (factor)}

\item{name.var}{name of variable}

\item{na.char}{a character vector of strings which are to be interpreted as NA values. Blank fields are also considered to be missing values in logical, integer, numeric and complex fields. Note that the test happens after white space is stripped from the input, so na.strings values may need their own white space stripped in advance.}
}
\description{
Crea un diccionario para los niveles de una variable factor
}
\examples{
\dontrun{
dc_fumador <- create_dictionary(var = factorDat$Fumador)
}
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{dictionary}
\keyword{factor}
\keyword{variable}
