% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.create.R
\name{date.create}
\alias{date.create}
\title{A date.create Function}
\usage{
date.create(dy, mth, yr)
}
\arguments{
\item{d}{vector numérico que incluya los dias.}

\item{m}{vector numérico que incluya los meses.}

\item{y}{vector numérico que incluya los años.}
}
\description{
Convierte tres vectores dia, mes, año en un fechas. En el caso de que falte el mes y o el día se indica el 6 y 15 respectivamente.
}
\examples{
day <- c(NA, NA, NA, 12, 20, NA, NA, NA, 19, NA, 25)
month <- c(NA, NA, 12, 7, 4, NA, 1, NA, 11, NA, 2)
year <- c(2000, NA, 2003, 2012, 2012, 1991, 2013, 2008, 2007, 1994, 2011)
date.create(d = day, m = month, y = year)
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{date}
\keyword{day}
\keyword{month}
\keyword{year}
