% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_ggplot.R
\name{desc_ggplot}
\alias{desc_ggplot}
\title{Genera gráficos descriptivos con ggplot2}
\usage{
desc_ggplot(
  dat,
  covariates = NULL,
  frml = NULL,
  y = NULL,
  nameFile = "descriptive_plots.pdf",
  topdf = FALSE,
  list.plots = FALSE,
  color = "#8D4ABA",
  rowcol = c(1, 1),
  show.freq = TRUE,
  bw = TRUE,
  size.n = 3,
  size.freq = 3,
  size.title = 10,
  show.pval = FALSE,
  show.n = TRUE,
  show.na = FALSE,
  ...
)
}
\arguments{
\item{dat}{Un data frame con los datos a analizar.}

\item{covariates}{Vector de nombres de las variables explicativas a analizar. Si es NULL, se analizan todas las variables.}

\item{frml}{Una fórmula opcional para especificar la variable dependiente y las covariables.}

\item{y}{Nombre de la variable dependiente (respuesta).}

\item{nameFile}{Nombre del archivo PDF donde se guardarán los gráficos (por defecto, "descriptive_plots.pdf").}

\item{topdf}{Lógico. Si es TRUE, guarda los gráficos en un archivo PDF.}

\item{list.plots}{Lógico. Si es TRUE, devuelve una lista con los gráficos en lugar de imprimirlos.}

\item{color}{Color principal para los gráficos (por defecto, "#8D4ABA").}

\item{rowcol}{Vector de longitud 2 indicando el número de filas y columnas en el PDF de salida.}

\item{show.freq}{Lógico. Si es TRUE, muestra la frecuencia en los gráficos de barras.}

\item{bw}{Lógico. Si es TRUE, agrega dispersión a los diagramas de caja.}

\item{size.n}{Tamaño del texto para la cantidad de observaciones.}

\item{size.freq}{Tamaño del texto para las frecuencias en los gráficos de barras.}

\item{size.title}{Tamaño del texto para los títulos de los gráficos.}

\item{show.pval}{Lógico. Si es TRUE, muestra los valores p en los gráficos bivariados.}

\item{show.n}{Lógico. Si es TRUE, muestra el número de observaciones en cada gráfico.}

\item{show.na}{Lógico. Si es TRUE, incluye valores NA en los gráficos.}
}
\value{
Si `list.plots` es TRUE, devuelve una lista con los gráficos generados. Si `topdf` es TRUE, guarda los gráficos en un archivo PDF.
        Si ambos son FALSE, imprime los gráficos en la consola.
}
\description{
Esta función crea gráficos descriptivos univariados y bivariados en función de las variables proporcionadas.
Se pueden generar histogramas, gráficos de barras, diagramas de caja y otras representaciones gráficas.
}
\examples{
# Generar gráficos descriptivos para todas las variables
desc_ggplot(mtcars)

# Especificar variables de análisis
desc_ggplot(mtcars, covariates = c("mpg", "hp"))

# Guardar gráficos en un PDF
#desc_ggplot(mtcars, topdf = TRUE, nameFile = "graficos.pdf")
}
