% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_lasso.R
\name{desc_lasso}
\alias{desc_lasso}
\title{A desc_lasso Function}
\usage{
desc_lasso(glmmod, s, y = NULL, data, show.all = FALSE)
}
\arguments{
\item{glmmod}{a fitted object of class inheriting from lasso}

\item{s}{Value(s) of the penalty parameter lambda at which predictions are required. Default is the entire sequence used to create the model.}

\item{y}{character variable. Outcome variable name}

\item{data}{data frame containing the outcome variable}

\item{show.all}{TRUE or FALSE. show all coefficients. Default value is dalse}
}
\description{
Genera resumen para los coeficientes de laso distintos de 0 ( o todos)
}
\keyword{lasso}
