% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descPlot.R
\name{descPlot}
\alias{descPlot}
\title{A descPlot Function}
\usage{
descPlot(dat, y = NULL, nameFile = "descriptive_plots.pdf",
  topdf = FALSE, subtitle = NULL, color = "#8D4ABA", rowcol = c(3,
  2), show.lg = FALSE, show.freq = TRUE, cex.lab = 1,
  cex.lg = 0.65, cex.main = 0.9, las = 0, do.test = FALSE,
  at.text = 1)
}
\arguments{
\item{dat}{data frame que contiene las variables a graficar.}

\item{nameFile}{nombre del fichero (tipo caracter) donde guardar los gr치ficos. Por defecto 'descriptive_plots.pdf'.}

\item{topdf}{valor l칩gico indicando si se quieren guardar los gr치ficos en pdf. Por defecto FALSE.}

\item{subtitle}{subtitol}

\item{color}{nombre del color para pintar el gr치fico}

\item{rowcol}{c(nrows, ncols) to create a matrix of nrows x ncols plots that are filled in by row}

\item{show.lg}{TRUE o FALSE indica si se muestra la leyenda. Por defecto FALSE.}

\item{show.freq}{TRUE o FALSE indica si se muestran las frecuencias. Por defecto TRUE}

\item{cex.lab}{expansion factor for axis names (bar labels). (size x labels)}

\item{cex.lg}{character expansion factor relative to current par('cex'). Used for text, and provides the default for pt.cex and title.cex.}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. 0: always parallel to the axis [default],
1: always horizontal, 2:always perpendicular to the axis, 3: always vertical.}

\item{do.test}{logical value si se quiere realizar test kruskall Wallis.}

\item{at.text}{if do.test TRUE, give location of each string in user coordinates. If the component of at corresponding to a particular text item is not a finite value (the default), the location will be determined by adj.}
}
\description{
Genera graficos univariantes para todas las variables que se indiquen, el formato de entrada es 'data.frame'
}
\examples{
df <- data.frame(Y=as.factor(rbinom(50,1,.40)),Y1=as.factor(rbinom(50,1,.40)),X = rnorm(50,10,1))
descPlot(dat = df, color = 'red', rowcol = c(1,2))
descPlot(dat = df, y = 'Y', color = 'red', rowcol = c(1,1), las = 2)
descPlot(mtc_bis)
descPlot(dat = mtc_bis, y ='gear', rowcol = c(2,2), do.test = TRUE, las = 2 )
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{descriptive}
\keyword{plots}
