% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.test.R
\name{diag.test}
\alias{diag.test}
\title{A diag.test Function}
\usage{
diag.test(pred, y, tag.healthy = levels(y)[1], nround = 2)
}
\arguments{
\item{y}{a  character string with the name of the variable that distinguishes healthy from diseased individuals}

\item{x}{either a character string with the name of the diagnostic test variable. (Potser una variable numerica o per exemple, una probabilitat de un model de regressio logistica)}

\item{frml}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. Es necesario usar este parametro cuando tengamos mas de una variable explicativa.}
}
\value{
variable:  taula detallada amb totes les mesures de classificació

summary: taula detallada amb totes les mesures de classificació per a cada una de les variables
}
\description{
mesures de clasificació
}
\examples{
data(aSAH)
y <- aSAH$outcome
p1 <- factor(ifelse(aSAH$s100b >= 0.22, levels(aSAH$outcome)[2], levels(aSAH$outcome)[1]  ))
diag.test(y = y, pred = p1)

df_pred <- data.frame(p1 = factor(ifelse(aSAH$s100b >= 0.22, levels(aSAH$outcome)[2], levels(aSAH$outcome)[1]  )),
p2 = factor(ifelse(aSAH$s100b >= 0.3, levels(aSAH$outcome)[2], levels(aSAH$outcome)[1]  )))
diag.test(y = y, pred = df_pred)
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{glm}
\keyword{roc}
\keyword{test}
