% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{A freq_table Function}
\usage{
freq_table(var.cat, y, data, margin = NULL, nround = 2, latex = TRUE,
  sz.tab = "footnotesize", title = "titles for the table")
}
\arguments{
\item{var.cat}{vector con el nombre de las variables categoricas a resumir}

\item{y}{nombre de la variable factor principal.}

\item{data}{data frame que contiene las variables a analizar}

\item{margin}{index, or vector of indices to generate margin for}

\item{nround}{integer indicating the number of decimal places (round).}

\item{latex}{TRUE o FALSE, para obtener tabla en formato .tex}

\item{sz.tab}{A character vector that is inserted just before the tabular environment starts. This can be used to set the font size and a variety of other table settings. Initial backslashes are automatically prefixed, if not supplied by user. Default value is "footnotesize"}

\item{title}{titles for the table}
}
\description{
Tabla de frecuencias relativas, por fila columnas o global.
}
\examples{
freq_table(var.cat = c("cyl","vs","gear", "carb"), y = "am", data = mtc_bis, latex = TRUE)
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{compare}
\keyword{freq}
\keyword{qualitative}
\keyword{tables}
