% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.uni.R
\name{glm.uni}
\alias{glm.uni}
\title{A glm.uni Function}
\usage{
glm.uni(y, var2test, data, size = 8.5, format = "latex",
  caption = "Univariate logistic regression", show.n = TRUE, group = TRUE)
}
\arguments{
\item{y}{response variable. Variable factor con 2 categorias}

\item{var2test}{nombre de las variables a testar mediante regresión logítica}

\item{size}{A numeric input for table font size}

\item{format}{a character string; possible values are latex, html, markdown, pandoc, and rst; this will be automatically determined if the function is called within knitr; it can also be set in the global option knitr.table.format; if format is a function, it must return a character string}

\item{caption}{Character vector containing the table's caption or title. Default value is "Univariate logistic regression"}

\item{show.n}{TRUE o FALSE muestra el total de individuos usados para el ajuste del modelo. Default value is "TRUE".}

\item{group}{TRUE o FALSE mostrar variables agrupadas en la tabla}
}
\description{
Genera tabla con los coeficientes OR , intervalos de confianza y p-valores de un modelo de regresión logística
}
\examples{
df <- data.frame( score = rnorm(50,10,1), hores = rnorm(50,10,1), mort = as.factor(rbinom(50,1,.40)) )
glm.uni(y = "mort", var2test = c("score", "hores"), data = df, format = "latex", size = 10)
}
\keyword{OR}
\keyword{logistic}
\keyword{regresion}
\keyword{summary}
