% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installifnot.R
\name{installBiocifnot}
\alias{installBiocifnot}
\title{installifnot Function}
\usage{
installBiocifnot(pckgName)
}
\arguments{
\item{pkg}{Nombre del paquete en formato caracter}
}
\description{
Comprueba si un paquete de Bioconductor esta instalado y si no lo instala y carga.
}
\examples{
# installBiocifnot("CMA")
}
\keyword{install}

