% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installifnot.R
\name{installGitifnot}
\alias{installGitifnot}
\title{installifnot Function}
\usage{
installGitifnot(pathGit, pckgName)
}
\arguments{
\item{pathGit}{carpeta on esta el paquet dins de git amb "/" final}

\item{pkg}{Nombre del paquete en formato caracter}
}
\description{
Comprueba si un paquete de Github esta instalado y si no lo instala y carga.
}
\examples{
# installGitifnot("miriammota/","mmotaF")
}
\keyword{install}

