% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installifnot.R
\name{installifnot}
\alias{installifnot}
\title{installifnot Function}
\usage{
installifnot(pckgName)
}
\arguments{
\item{pkg}{Nombre del paquete en formato caracter}
}
\description{
Comprueba si un paquete de R cran esta instalado y si no lo instala y carga.
}
\examples{
# installifnot("rpart")
}
\keyword{install}

