% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_surv_table.R
\name{make_surv_table}
\alias{make_surv_table}
\title{Create a nice data frame from a survival model for automatically generating a table.}
\usage{
make_surv_table(fit_surv, nround = 2, event = FALSE)
}
\arguments{
\item{nround}{integer indicating the number of decimal places (round) or significant digits (signif) to be used. Negative values are allowed}

\item{event}{cumulative events (f(y) = 1-y)}

\item{survobj}{\code{\link[survival]{survfit}} object}
}
\value{
a data frame with labeled columns
}
\description{
Create a nice data frame from a survival model for automatically generating a table.
}
\examples{
data(aml)
surv_fit <-survfit(Surv(time = aml$time, event = aml$status) ~ 1)
stargazer(make_surv_table(surv_fit),
           summary = FALSE, type = c('latex'), title = 'Summary of status survival fit',
          rownames = FALSE)
}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{latex}
\keyword{survival}
\keyword{table}
