% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multGGplot.R
\name{multGGplot}
\alias{multGGplot}
\title{A multGGplot Function}
\usage{
multGGplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}

\item{plotslist}{plots separados por coma}
}
\description{
Similar a par(mfrow=c(row,col)) para "ggplot" (Obtenida de cookbook for R )
}
\examples{
p1 <- qplot(gear, mpg, data=mtcars, geom=c("boxplot", "jitter"),fill=gear)
p2 <- p3 <- p4 <- p1
multGGplot(p1,p2,p3,p4, cols=2)
}
\keyword{multiGGplot}

