% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postTexCG.R
\name{postTexCG}
\alias{postTexCG}
\title{A postTexCG Function}
\usage{
postTexCG(nameFile)
}
\arguments{
\item{nameFile}{nombre del archivo con ruta incluida}
}
\value{
Reescribe el fichero eliminando "2*"
}
\description{
Corrige problema con funcion export2latex del paquete compareGroups, 
donde por defecto indica: "2*p.overall" y "2*N"
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{compareGroups}
\keyword{error}
\keyword{export2latex}
\keyword{tex}

