% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_name.R
\name{re_name}
\alias{re_name}
\title{A re_name Function}
\usage{
re_name(data)
}
\arguments{
\item{dat}{data frame que contiene las variables a evaluar}
}
\description{
Renombra los identificadores de columna, substituyendo '.' por '_', eliminando acentos, ec.
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{correct}
\keyword{dataframe}
\keyword{names}
