% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_factor.R
\name{recode_factor}
\alias{recode_factor}
\title{A recode_factor Function}
\usage{
recode_factor(dict, var, na.char = "NA")
}
\arguments{
\item{dict}{data.frame con dos columnas, columna 1,rawtext, niveles posibles; columna 2,recoded, niveles correctos(an object of type create_dictionary.)}

\item{var}{object to be coerced or tested. (factor)}

\item{na.char}{caracter que sera interpretado como NA}
}
\description{
Crea un diccionario para los niveles de una variable factor
}
\examples{
\dontrun{
dc_fumador <- create_dictionary(factorDat$Fumador)
table(factorDat$Fumador)
factorDat$Fumador <- recode_factor(dict = dc_fumador, var = factorDat$Fumador)
table(factorDat$Fumador)
}
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{clean}
\keyword{dictionary}
\keyword{factor}
\keyword{matching}
\keyword{variable}
