% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.var.labels.R
\name{set.var.labels}
\alias{set.var.labels}
\title{A set.var.labels Function}
\usage{
set.var.labels(dataframe, label.vector)
}
\arguments{
\item{dataframe}{en el cual queremos añadir etiquetas}

\item{label.vector}{vector string con el nombre de las etiquetas de las variables}
}
\description{
Preparar etiquetas para las variables
}
\examples{
#dat <- data.frame(a=c(0,1,2),b=c(1,3,4),d=c(3,4,5))
#labels <- c("Letter A","Letter B", "Letter C") ## vector con etiquetes
#var.labels <- set.var.labels (dat,as.character(labels))
#label(dat) <- lapply(names(var.labels), function(x) label(dat[,x]) = var.labels[x])
}
\keyword{labels}

