% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_bw.R
\name{boxplot_bw}
\alias{boxplot_bw}
\title{A boxplot_bw Function}
\usage{
boxplot_bw(y, group = NULL, frml = NULL, dat, las = 0,
  title.plot = NULL, sub.plot = NULL, ylim.plot = NULL,
  cex.lab = 1, cex.main = 1, cex.pval = 0.6, cex.n = 0.5,
  ylab = "", xlab = NULL, do.test = FALSE, color = NULL)
}
\arguments{
\item{y}{name numeric vector of data values.}

\item{group}{name factor vector. Default value is NULL.}

\item{frml}{Right side of ~ must have the terms in an additive way, and left side of ~ must contain the name of the grouping variable or can be left in blank (in this latter case descriptives for whole sample are calculated and no test is performed).}

\item{dat}{matrix or data frame containing the variables in the formula.}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. 0: always parallel to the axis [default],
1: always horizontal, 2:always perpendicular to the axis, 3: always vertical.}

\item{title.plot}{a main title for the plot}

\item{sub.plot}{a sub title for the plot}

\item{ylim.plot}{is vector which contains lower and upper limits which are to appear on the y axes.}

\item{cex.lab}{size of the axis label text with a numeric value.}

\item{cex.main}{expansion factor for main names (size main)}

\item{cex.pval}{size of the p-value text with a numeric value.}

\item{cex.n}{expansion factor for size n}

\item{ylab}{a title for the y axis}

\item{xlab}{a title for the x axis.}

\item{do.test}{logical value si se quiere realizar test kruskall Wallis.}
}
\description{
Boxplot incluyendo puntos individuales
}
\examples{
df <- data.frame(runif = c(runif(100, min = -3, max = 3),
rep(0,25)), rnorm = c(rnorm(100),rep(0,25)) )
boxplot_bw(dat = df, y = 'rnorm' )
boxplot_bw(dat = mtc_bis, y = 'qsec', title.plot = "Title" )
boxplot_bw(dat = mtc_bis, y = 'qsec', group = 'gear',
title.plot = "Boxplot per grup", do.test = TRUE, las = 2)
boxplot_bw(dat = df, frml =   ~ rnorm )
boxplot_bw(dat = mtc_bis, , frml =  gear ~ qsec ,
title.plot = "Boxplot per grup", do.test = TRUE, las = 2)
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{boxplot}
\keyword{descriptive}
\keyword{plots}
