% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_bw.R
\name{boxplot_bw}
\alias{boxplot_bw}
\title{titulo del grafico}
\usage{
boxplot_bw(y, x, dat, las = 0, title = NULL, ylim = NULL)
}
\arguments{
\item{y}{nombre de la variable factor}

\item{x}{nombre de la variable numerica}

\item{dat}{data frame que contiene las variables a graficar.}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. 0: always parallel to the axis [default],
1: always horizontal, 2:always perpendicular to the axis, 3: always vertical.}

\item{ylim}{ylim}
}
\description{
A boxplot_bw Function
}
\details{
Boxplot incluyendo puntos individuales
}
\examples{
boxplot_bw(dat = mtc_bis, x = "qsec" ,y ="gear")
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{boxplot}
\keyword{descriptive}
\keyword{plots}
