% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_ggplot.R
\name{desc_ggplot}
\alias{desc_ggplot}
\title{Genera gráficos descriptivos con ggplot2}
\usage{
desc_ggplot(
  dat,
  covariates = NULL,
  frml = NULL,
  y = NULL,
  nameFile = "descriptive_plots.pdf",
  topdf = FALSE,
  list.plots = FALSE,
  color = "#8D4ABA",
  rowcol = c(1, 1),
  show.freq = TRUE,
  bw = TRUE,
  size.n = 3,
  size.freq = 2.5,
  size.title = 10,
  size.pval = 3,
  show.pval = FALSE,
  show.n = TRUE,
  show.na = FALSE,
  legend.position = "right",
  ...
)
}
\arguments{
\item{dat}{Data frame con los datos a analizar.}

\item{covariates}{Vector de nombres de variables a analizar. Si es NULL, se analizan todas las variables.}

\item{frml}{Fórmula opcional para especificar la variable dependiente y las covariables.}

\item{y}{Nombre de la variable dependiente (opcional para análisis bivariado).}

\item{nameFile}{Nombre del archivo PDF donde se guardarán los gráficos (si `topdf = TRUE`).}

\item{topdf}{Lógico. Si es TRUE, guarda los gráficos en un archivo PDF.}

\item{list.plots}{Lógico. Si es TRUE, devuelve una lista con los gráficos generados.}

\item{color}{Color para los gráficos de variables numéricas.}

\item{rowcol}{Vector de longitud 2 especificando el número de filas y columnas en la salida de gráficos.}

\item{show.freq}{Lógico. Si es TRUE, muestra frecuencias en gráficos de barras.}

\item{bw}{Lógico. Si es TRUE, agrega dispersión en gráficos de cajas.}

\item{size.n}{Tamaño del texto para `n` en los gráficos.}

\item{size.freq}{Tamaño del texto para frecuencias en gráficos de barras.}

\item{size.title}{Tamaño del título de los gráficos.}

\item{size.pval}{Tamaño del texto del p-valor en gráficos bivariantes.}

\item{show.pval}{Lógico. Si es TRUE, muestra el p-valor en gráficos bivariantes.}

\item{show.n}{Lógico. Si es TRUE, muestra el número de observaciones en los gráficos.}

\item{show.na}{Lógico. Si es TRUE, incluye valores NA en los gráficos.}

\item{legend.position}{Posición de la leyenda en gráficos bivariantes.}
}
\value{
Si `list.plots` es TRUE, devuelve una lista con los gráficos generados. Si `topdf` es TRUE, guarda los gráficos en un archivo PDF.
        Si ambos son FALSE, imprime los gráficos en la consola.
}
\description{
Esta función crea gráficos descriptivos univariados y bivariados en función de las variables proporcionadas.
Se pueden generar histogramas, gráficos de barras, diagramas de caja y otras representaciones gráficas.
}
\examples{
# Generar gráficos descriptivos para todas las variables
#desc_ggplot(mtcars)

# Especificar variables de análisis
#desc_ggplot(mtcars, covariates = c("mpg", "hp"))

# Guardar gráficos en un PDF
#desc_ggplot(mtcars, topdf = TRUE, nameFile = "graficos.pdf")
}
