% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descPlot.R
\name{descPlot}
\alias{descPlot}
\title{A descPlot Function}
\usage{
descPlot(dat, y = NULL, nameFile = "descriptive_plots.pdf", topdf = FALSE,
  subtitle = NULL, color = "#8D4ABA", rowcol = c(3, 2), show.lg = FALSE,
  show.freq = TRUE, cex.lab = 0.8, las = 3)
}
\arguments{
\item{dat}{data frame que contiene las variables a graficar.}

\item{nameFile}{nombre del fichero (tipo caracter) donde guardar los gr치ficos. Por defecto "descriptive_plots.pdf".}

\item{topdf}{valor l칩gico indicando si se quieren guardar los gr치ficos en pdf. Por defecto FALSE.}

\item{subtitle}{subtitol}

\item{color}{nombre del color para pintar el gr치fico}

\item{rowcol}{c(nrows, ncols) to create a matrix of nrows x ncols plots that are filled in by row}

\item{show.lg}{TRUE o FALSE indica si se muestra la leyenda. Por defecto FALSE.}

\item{show.freq}{TRUE o FALSE indica si se muestran las frecuencias. Por defecto TRUE}

\item{cex.y}{size labels variable resposta}
}
\description{
Genera graficos univariantes para todas las variables que se indiquen, el formato de entrada es "data.frame"
}
\examples{
df <- data.frame(Y=as.factor(rbinom(50,1,.40)),X = rnorm(50,10,1))
descPlot(dat = df, color = "red", rowcol = c(1,2))
descPlot(dat = df, y = "Y", color = "red", rowcol = c(1,1))
descPlot(mtc_bis)
descPlot(dat = mtc_bis, y ="gear", rowcol = c(2,2))
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{descriptive}
\keyword{plots}
