% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descPlot.R
\name{descPlot}
\alias{descPlot}
\title{A descPlot Function}
\usage{
descPlot(dat, topdf = FALSE, nameFile = "descriptive_plots.pdf")
}
\arguments{
\item{dat}{data frame que contiene las variables a graficar.}

\item{topdf}{valor lógico indicando si se quieren guardar los gráficos en pdf. Por defecto FALSE.}

\item{nameFile}{nombre del fichero (tipo caracter) donde guardar los gráficos. Por defecto "descriptive_plots.pdf".}
}
\description{
Genera graficos univariantes para todas las variables que se indiquen, el formato de entrada es "data.frame"
}
\examples{
df <- data.frame(y = as.factor(c(0,0,0,0,0,0,1,1,1,1,1,1,0,0,0,0,0,1,1,1,1,1)),X = c(20,2,1,2,23,3,4,5,15,-5,2,1,3,10,4,87,5,3,7,0,1,-2))
descPlot(dat = df)
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{descriptive}
\keyword{plots}

