% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_list.R
\name{dictionary_list}
\alias{dictionary_list}
\title{A dictionary_list Function}
\usage{
dictionary_list(variables, data)
}
\arguments{
\item{var}{nombre de las variables factor a recodificar}

\item{dat}{data frame, list or environment (or object coercible by 'as.data.frame' to a data frame) containing the variables}
}
\description{
Crea una lista de diccionarios para los niveles de una seria de variables factor
}
\examples{
# dictionary_all <- dictionary_list(names(factorDat), factorDat)
# Asignar nuevos niveles a todas las variables
# for (i in 1:length(dictionary_all)) {
# factorDat[,gsub("^dc_","",names(dictionary_all)[i])] <- recode_factor(dict = dictionary_all[[i]],
#                                                                        var = factorDat[,gsub("^dc_","",names(dictionary_all)[i])] )
#                                                                        }
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{dictionary}
\keyword{factor}
\keyword{variable}
