% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_list.R
\name{dictionary_list}
\alias{dictionary_list}
\title{A dictionary_list Function}
\usage{
dictionary_list(variables, data)
}
\arguments{
\item{var}{nombre de las variables factor a recodificar}

\item{dat}{data frame, list or environment (or object coercible by 'as.data.frame' to a data frame) containing the variables}
}
\description{
Crea una lista de diccionarios para los niveles de una seria de variables factor
}
\examples{
\dontrun{
if (!file.exists('dades/dictionaries.rda')) {
dictionaries <- dictionary_list(variables = names(factorDat), data = factorDat)
save(dictionaries, file = 'dades/dictionaries.rda')
}else{
   load('dades/dictionaries.rda')
  }
Asignar nuevos niveles a todas las variables
for (i in 1:length(dictionaries)) {
factorDat[,gsub('^dc_','',names(dictionaries)[i])] <- recode_factor(dict = dictionaries[[i]],
var = factorDat[,gsub('^dc_','',names(dictionaries)[i])] )
 }
}
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{dictionary}
\keyword{factor}
\keyword{variable}
