% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doROC.R
\name{doROC}
\alias{doROC}
\title{A doROC Function}
\usage{
doROC(frml, dat, titol, validation = FALSE, test, test_y,
  col.thres = "blue", col.ic = "#aaddddAA", x.axes = FALSE)
}
\arguments{
\item{frml}{objeto tipo formula indicando como variable respuesta una variable binaria y como explicativa una contínua}

\item{dat}{data.frame donde estan las variables de la formula.}

\item{titol}{tipo caracter. Título para el gráfico}

\item{validation}{valor lógico indicando si se entreran datos externos para testar el punto de corte elegido en el grafico}

\item{test}{data.frame indicando los nuevos valores de la variable explicativa}

\item{test_y}{vector factor indicando grupo de los nuevos individuos}

\item{col.thres}{color de la cruz que indica el punto de corte óptimo en el gráfico}

\item{col.ic}{color para el intervalo de confianza. Debe ser translucido, por lo que se puede usar la función makeTransparent para cualquier color R.}

\item{x.axes}{a logical indicating if the specificity axis (x axis) must be plotted as as decreasing “specificity” (FALSE, the default) or increasing “1 - specificity” (TRUE) as in most legacy software. This affects only the axis, not the plot coordinates.}
}
\value{
auc: Area bajo la curva y correspondiente intervalo de confianza

pvalue: p-valor de la variable explicativa en el modelo de regresion logística

thres.best: punto de corte  óptimo calculado con el estad ́ıstico de Youden

misc.test: en el caso de tener datos de validación, la tasa de error de clasificacion utilizando como punto de corte el thres.best.
}
\description{
Creación de curvas ROC con regresión logística mediante el paquete pROC y validación con datos externos.
Para realizar las curvas ROC, en primer lugar, se modeliza - para el conjunto de datos “training set”- la variable respuesta grupo 
mediante regresión logística teniendo en cuenta la variable explicativa. Una vez ajustado el modelo se realiza las curva
ROC(representación gráfica de la sensibilidad en frente a la especificidad).
}
\examples{
y <- as.factor(rbinom(50,1,.40))
x <- rnorm(50,10,1)
resROC <- doROC (frml = y ~ x, titol = "prova",
                 validation = TRUE,
                 test = data.frame(x=c(1,2,3)),
                 test_y = as.factor(c(0,1,1)))
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{glm}
\keyword{roc}
\keyword{test}
\keyword{validation}

