% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doROC.R
\name{doROC}
\alias{doROC}
\title{A doROC Function}
\usage{
doROC(x, group, frml, dat, tag.healthy = NULL, title = NULL,
  doPlot = TRUE, cex.main = 0.9, show.lg = TRUE, show.cascon = TRUE,
  show.detail = TRUE, xtab = FALSE, modGLM = NULL, direction = c("<",
  ">"), ...)
}
\arguments{
\item{x}{either a character string with the name of the diagnostic test variable. (Potser una variable numerica o per exemple, una probabilitat de un model de regressio logistica)}

\item{frml}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. Es necesario usar este parametro cuando tengamos mas de una variable explicativa.}

\item{dat}{data frame containing the variables in the formula.}

\item{tag.healthy}{the value codifying healthy individuals in the status variable. Por defecto nivel de referencia levels(dat[,group])[1]}

\item{title}{a main title for the plot}

\item{cex.main}{expansion factor for main names (size main)}

\item{show.lg}{A logical value indicating whether show a legend Default value is TRUE}

\item{show.cascon}{A logical value indicating whether show number cases/controls. Default value is TRUE}

\item{show.detail}{A logical value indicating whether show detail output. Default value is TRUE}

\item{xtab}{A logical value indicating whether the output is a xtable. Default value is FALSE.}

\item{modGLM}{Valor logico que indica si se realiza regresion logistica. En el caso de indicar TRUE, el 'thres.best' indicara el punto de corte como probabilidad de predicción. En el caso de indicar FALSE, 'cutoff' nos indicara el punto de corte real en la variable. Es necesario indicar TRUE cuando querramos evaluar mas de una variable.}

\item{direction}{character string specifying the direction to compute the ROC curve. By default individuals with a test value lower than the cutoff are classified as healthy (negative test), whereas patients with a test value greater than (or equal to) the cutoff are classified as diseased (positive test). If this is not the case, however, and the high values are related to health, this argument should be established at ">".}

\item{y}{a  character string with the name of the variable that distinguishes healthy from diseased individuals}

\item{doPlotA}{logical value indicating whether show a plot. Default value is TRUE}
}
\value{
auc: Area bajo la curva y correspondiente intervalo de confianza

cutoff.probability: en el caso de haber realizado modGLM, punto de corte óptimo de la probabilidad de predicción calculado con el indice de Youden

cutoff.variable: punto de corte de la variable cuantitativa. SOLO CUANDO EVALUAMOS UNA UNICA VARIABLE CUANTITATIVA.

youden: the optimal value of the method considered for selecting the optimal cutpoint, i.e., the value of the criterion at the optimal cutpoint.

res_detail: taula detallada amb totes les sensibilitats i especificitats

res_sum:  the optimal cutpoint(s) obtained with the method(s) selected; its/their accuracy measures and the area under ROC curve (AUC)
}
\description{
Creación de curvas ROC (representación gráfica de la sensibilidad en frente a la especificidad) con o sin regresión logística mediante el paquete pROC y optimalCutpoints.
Se calcula el punto de corte optimo mediante youden y se obtienen las medidas de clasificacion.
}
\examples{
set.seed(12)
df <- data.frame(y = as.factor(rbinom(50,1,.40)),x = rnorm(50,10,1))
resROC <- doROC (frml = y ~ x, title = 'prova1', cex.main = 0.6, dat = df, modGLM = FALSE, direction = ">")
resROC$cutoff.variable
resROC <- doROC (frml = y ~ x, title = 'prova1', cex.main = 0.6, dat = df, modGLM = TRUE)
# si usamos el parametro modGLM = TRUE y queremos obtener el punto de corte real en la variable.
# Esto SOLO funciona si tenemos unicamente UNA variable explicativa.
resROC$cutoff.variable
(pt <- resROC$dat$x[which(resROC$dat$pred == resROC$cutoff.probability)])

}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{glm}
\keyword{roc}
\keyword{test}
