% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_surv_table.R
\name{make_surv_table}
\alias{make_surv_table}
\title{Create a nice data frame from a survival model for automatically generating a table.}
\usage{
make_surv_table(survobj, timevec, scale = 1, col_names = NULL)
}
\arguments{
\item{survobj}{\code{\link[survival]{survfit}} object}

\item{timevec}{a vector of times for survival summary}

\item{scale}{a number that scales the timevec, e.g., if your data is recorded in days and you want months displayed, use scale = 30}
}
\value{
a data frame with labeled columns
}
\description{
Create a nice data frame from a survival model for automatically generating a table.
}
\examples{
data(aml)
surv_fit <-survfit(Surv(time = aml$time, event = aml$status) ~ 1)
stargazer(make_surv_table(surv_fit, summary(surv_fit)$time),
           summary = FALSE, type = c('latex'), title = 'Summary of HMOHIV survival fit',
          rownames = FALSE)
}
\keyword{latex}
\keyword{survival}
\keyword{table}
