% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_mod.R
\name{tabOR_lr}
\alias{tabOR_lr}
\title{A desc_mod Function}
\usage{
tabOR_lr(...)
}
\arguments{
\item{mod}{a fitted object of class inheriting from "glm", "clogit","cox" or "lm".}

\item{xtab}{TRUE o FALSE, para obtener tabla en formato .tex}

\item{title}{if xtab = T, Character vector containing the table's caption or title.}

\item{xtab.type}{Type of table to produce. Possible values for type are "latex" or "html". Default value is "latex".}

\item{sz.latex}{A character vector that is inserted just before the tabular environment starts. This can be used to set the font size and a variety of other table settings. Initial backslashes are automatically prefixed, if not supplied by user. Default value is "small".}

\item{label}{Character vector of length 1 containing the LaTeX label. Default value is NULL.}

\item{show.intcp}{TRUE o FALSE, indica si se muestra o no el intercept del modelo. En ambos casos el modelo se ha calcula con intercept. Default value is "FALSE".}

\item{show.n}{TRUE o FALSE muestra el total de individuos usados para el ajuste del modelo. Default value is "TRUE".}

\item{show.aov.pval}{TRUE o FALSE muestra el p-valor del modelo global. Default value is "TRUE".}

\item{show.pretty}{TRUE o FALSE muestra las 'labels' de las variables. Solo funciona para lm y glm . Default value is "FALSE".}

\item{group_rw}{TRUE o FALSE  agrupa las filas por variables. Default value is "FALSE".}
}
\description{
Genera tabla resumen para modelos lineales,logisticos y cox
}
\examples{

}
\keyword{OR}
\keyword{logistica}
\keyword{regresion}
