% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabOR_lr.R
\name{tabOR_lr}
\alias{tabOR_lr}
\title{A tabOR_lr Function}
\usage{
tabOR_lr(mod, xtab = FALSE, title = "title", xtab.type = "latex",
  sz.latex = "small", label = NULL, show.intcp = FALSE, show.n = TRUE,
  show.aov.pval = TRUE)
}
\arguments{
\item{mod}{a fitted object of class inheriting from "glm".}

\item{xtab}{TRUE o FALSE, para obtener tabla en formato .tex}

\item{title}{if xtab = T, Character vector containing the table's caption or title.}

\item{xtab.type}{Type of table to produce. Possible values for type are "latex" or "html". Default value is "latex".}

\item{sz.latex}{A character vector that is inserted just before the tabular environment starts. This can be used to set the font size and a variety of other table settings. Initial backslashes are automatically prefixed, if not supplied by user. Default value is "small".}

\item{label}{Character vector of length 1 containing the LaTeX label. Default value is NULL.}

\item{show.intcp}{TRUE o FALSE, indica si se muestra o no el intercept del modelo. En ambos casos el modelo se ha calcula con intercept. Default value is "FALSE".}

\item{show.n}{TRUE o FALSE muestra el total de individuos usados para el ajuste del modelo. Default value is "TRUE".}

\item{show.aov.pval}{TRUE o FALSE muestra el p-valor del modelo global. Default value is "TRUE".}
}
\description{
Genera tabla con los coeficientes OR , intervalos de confianza y p-valores de un modelo de regresión logística
}
\examples{
df <- data.frame( x = rnorm(50,10,1), y = as.factor(rbinom(50,1,.40)) )
mod <- glm(y ~ x, data = df, family = binomial)
tabOR_lr(mod, xtab = FALSE,title='OR de los coeficientes')
}
\keyword{OR}
\keyword{logistica}
\keyword{regresion}
