% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabOR_lr.R
\name{tabOR_lr}
\alias{tabOR_lr}
\title{A tabOR_lr Function}
\usage{
tabOR_lr(mod, xtab = FALSE, title = "title")
}
\arguments{
\item{mod}{Modelo de regresión logística ("glm")}

\item{xtab}{TRUE o FALSE, para obtener tabla en formato .tex}

\item{title}{Solo en el caso de xtab = TRUE. Cabecera de la tabla.}
}
\description{
Genera tabla con los coeficientes OR , intervalos de confianza y p-valores de un modelo de regresión logística
}
\examples{
#tabOR_lr(glm(rnorm(50,10,1)~ as.factor(rbinom(50,1,.40))),
#xtab=TRUE,title="OR de los coeficientes")
}
\keyword{OR}
\keyword{logisitica}
\keyword{regresion}
